login api

+++ Assets/Scripts/Scriptable Object/Managers/UserManager.cs
@@ -31,12 +31,12 @@ namespace KWY
         private string _accountId; // °íÀ¯°ª
         private Sprite _userIcon;
         private int _userLevel;
-        private int _userId;
+        private ulong _userId;
 
         public static string AccountId { get { return Instance._accountId; } }
         public static Sprite UserIcon { get { return Instance._userIcon; } }
         public static int UserLevel { get { return Instance._userLevel; } }
-        public static int UserId { get { return Instance._userId; } }
+        public static ulong UserId { get { return Instance._userId; } }
 
         /// <summary>
         /// It should be called when login is successful; before using the instance; 
@@ -45,7 +45,7 @@ namespace KWY
         /// <param name="accountId">user unique id allocated from the server</param>
         /// <param name="userLevel">user level</param>
         /// <param name="userId">user id</param>
-        public static void InitData(Sprite userIcon, string accountId, int userLevel, int userId)
+        public static void InitData(Sprite userIcon, string accountId, int userLevel, ulong userId)
         {
             Instance._accountId = accountId;
             Instance._userIcon = userIcon;
